+++ src/openms/include/OpenMS/FORMAT/sources.cmake
@@ -36,6 +36,7 @@ ExperimentalDesignFile.h
 FASTAFile.h
 FeatureXMLFile.h
 FileHandler.h
+GNPSMGFFile.h
 GzipIfstream.h
 GzipInputStream.h
 HDF5Connector.h
