+++ src/tests/topp/GNPSExport_2_merged.ini
@@ -16,8 +16,7 @@
       <ITEM name="no_progress" value="false" type="bool" description="Disables progress logging to command line" required="false" advanced="true" />
       <ITEM name="force" value="false" type="bool" description="Overwrite tool specific checks." required="false" advanced="true" />
       <ITEM name="test" value="false" type="bool" description="Enables the test mode (needed for internal use only)" required="false" advanced="true" />
-      <NODE name="merged_spectra" description="Options for exporting mgf file with merged spectra per consensusElement">
-        <ITEM name="precursor_mass_tolerance" value="0.02" type="double" description="Precursor mass tolerance (Da) for ms annotations" required="false" advanced="false" />
+      <NODE name="merged_spectra" description="Options for exporting mgf file with merged spectra per consensusElement">      
         <ITEM name="cos_similarity" value="0.85" type="double" description="Cosine similarity threshold for merged_spectra output" required="false" advanced="false" />
       </NODE>
     </NODE>
