+++ src/tests/topp/GNPSExport_3_binsize.ini
@@ -17,7 +17,6 @@
       <ITEM name="force" value="false" type="bool" description="Overwrite tool specific checks." required="false" advanced="true" />
       <ITEM name="test" value="false" type="bool" description="Enables the test mode (needed for internal use only)" required="false" advanced="true" />
       <NODE name="merged_spectra" description="Options for exporting mgf file with merged spectra per consensusElement">
-        <ITEM name="precursor_mass_tolerance" value="0.02" type="double" description="Precursor mass tolerance (Da) for ms annotations" required="false" advanced="false" />
         <ITEM name="cos_similarity" value="0.85" type="double" description="Cosine similarity threshold for merged_spectra output" required="false" advanced="false" />
       </NODE>
     </NODE>
